Moves smartcore::neighbors::KNNAlgorithmName to smartcore::algorithm::neighbour::KNNAlgorithmName
