+++ src/neighbors/knn_classifier.rs
@@ -38,7 +38,8 @@ use crate::error::Failed;
 use crate::linalg::{row_iter, Matrix};
 use crate::math::distance::Distance;
 use crate::math::num::RealNumber;
-use crate::neighbors::{KNNAlgorithm, KNNAlgorithmName, KNNWeightFunction};
+use crate::neighbors::KNNWeightFunction;
+use crate::algorithm::neighbour::{KNNAlgorithm, KNNAlgorithmName};
 
 /// `KNNClassifier` parameters. Use `Default::default()` for default values.
 #[derive(Serialize, Deserialize, Debug)]
